Revert "#121 fix compile errors after cherry picking from gh121-webui branch"
This reverts commit 9f850579867eeb761797bbd72189f2092dc2b36a.

+++ metasfresh-webui-api/src/main/java/de/metas/ui/web/pattribute/ASIDescriptorFactory.java
@@ -204,7 +204,7 @@ public class ASIDescriptorFactory
 	{
 		final ASILayout.Builder layout = ASILayout.builder()
 				.setCaption(attributeSetDescriptor.getCaption())
-				.setM_AttributeSet_ID(attributeSetDescriptor.getDocumentTypeId().toInt());
+				.setM_AttributeSet_ID(attributeSetDescriptor.getDocumentTypeId());
 
 		attributeSetDescriptor.getFields()
 				.stream()
